+++ home/forms.py
@@ -78,35 +78,6 @@ class UserForm(forms.ModelForm):
         return cd['password2']
 
 
-class CommunityPartnerForm(forms.ModelForm):
-
-    class Meta:
-        model = CommunityPartner
-        fields = ('name', 'website_url', 'community_type', 'k12_level', 'address_line1', 'address_line2', 'county',
-                  'city', 'state', 'zip')
-        widgets = {'name': forms.TextInput({'placeholder': 'Community Partner Name'}),
-                   'website_url': forms.TextInput({'placeholder': 'https://www.unomaha.edu'}),
-                   'k12_level' :forms.TextInput({'placeholder': 'If your community type is K12, Please provide the k12-level'}),
-                   }
-        def clean_website_url(self):
-         data = self.cleaned_data.get('website_url')
-         if  not  data.startswith ('https://'):
-            raise forms.ValidationError('Url should start from "https://".')
-         return data
-
-
-class CommunityContactForm(forms.ModelForm):
-    class Meta:
-        model = Contact
-        fields = ('first_name',
-                  'last_name',
-                  'work_phone',
-                  'cell_phone',
-                  'email_id',
-                  'contact_type')
-        widgets = {'work_phone': forms.TextInput({'placeholder': '5714200002'}),
-                   'cell_phone': forms.TextInput({'placeholder': '5714200002'}),
-                  }
 
 
 class UploadProjectForm(forms.ModelForm):
@@ -196,19 +167,6 @@ class UploadDepartment(ModelForm):
     class Meta:
         model = Department
         fields = '__all__'
-
-
-class CommunityMissionForm(ModelForm):
-
-    mission_choices = (
-        ('Primary', 'Primary'),
-        ('Secondary', 'Secondary'),
-        ('Other', 'Other'),
-    )
-
-    class Meta:
-        model = CommunityPartnerMission
-        fields = ('mission_type' , 'mission_area')
