+++ home/templates/home/Campus_Home.html
@@ -72,10 +72,10 @@ color:black;
             <li class=""><a href="#">Hello, {{user.username}}</a></li>
             <ul class="navbar-nav ml-auto">
                 <li class="nav-item">
-                    <a class="nav-link" href='/communitypartnerproject/'><i class="fa fa-tachometer"></i>Dashboard</a>
+                    <a class="nav-link" href="#"><i class="fa fa-tachometer"></i>Dashboard</a>
                 </li>
                 <li class="nav-item">
-                    <a class="nav-link" href="#"><i class="fa fa-product-hunt"></i>Project</a>
+                    <a class="nav-link" href='/project/'><i class="fa fa-product-hunt"></i>Project</a>
                 </li>
                 <li class="nav-item">
 
@@ -156,13 +156,14 @@ color:black;
 
 <!--Contact Begins -->
 
-<div class="container">
-
-{% block content %}
+<{% block content %}
+    <div class="container">
+        <div class="col-8">
+          <h3>Welcome Campus Partner </h3>
+        </div>
+    </div>
 
 {% endblock %}
-
-</div>
 
 <section id="contact">
       <div class="container">
