+++ home/templates/home/Community_Home.html
@@ -116,7 +116,14 @@ color:black;
 <!-- Navbar ends-->
 
 <!-- Table Begins ---- Add Block Content Here-->
+{% block content %}
+    <div class="container">
+        <div class="col-8">
+          <h3>Welcome Community Partner </h3>
+        </div>
+    </div>
 
+{% endblock %}
 <div class="container" id="example1">
       <table id="example" class="table table-striped table-bordered dt-responsive nowrap" style="width:100%">
         <thead>
@@ -153,20 +160,13 @@ color:black;
 
 
 
-
-
-
-<div class="container">
-
-{% block content %}
-    <p>Welcome Community Partner </p>
-{% endblock %}
-
 </div>
 
 
 
 <!--Contact Begins -->
+
+
