+++ home/templates/home/base_home.html
@@ -66,7 +66,7 @@ margin-top:15px;
                     </a>
                 <div class="dropdown-menu" aria-labelledby="dropdown_target">
                     <a class="dropdown-item" href="{% url 'partners:registerCampusPartner' %}">Campus Partner</a>
-                    <a class="dropdown-item" href='/registerCommunityPartner/'>Community Partner</a>
+                    <a class="dropdown-item" href='/partners/registerCommunityPartner/'>Community Partner</a>
                 </div>
                 </li>
                 <li class="nav-item">
