+++ home/urls.py
@@ -14,7 +14,6 @@ urlpatterns = [
    path('registerCampusPartnerUser/', views.registerCampusPartnerUser, name='registerCampusPartnerUser'),
    path('registerCommunityPartnerUser/', views.registerCommunityPartnerUser, name='registerCommunityPartnerUser'),
    path('signupuser/registerCampusPartnerUser/', views.registerCampusPartnerUser, name='registerCampusPartnerUser'),
-   path('registerCommunityPartner/', views.registerCommunityPartner, name='registerCommunityPartner'),
    path('signupuser/registerCommunityPartnerUser/', views.registerCommunityPartnerUser,name='registerCommunityPartnerUser'),
    path('signup/', views.signup, name='signup'),
    path('signupuser/', views.signupuser, name='signupuser'),
@@ -22,4 +21,4 @@ urlpatterns = [
    path('upload_community/', views.upload_community, name='upload_community'),
    path('upload_campus/', views.upload_campus, name='upload_campus'),
 
-] + static(settings.MEDIA_URL, document_root=settings.MEDIA_ROOT)
+]
