+++ partners/templates/registration/campus_partner_register.html
@@ -61,7 +61,7 @@ font-weight: bold;
 </style>
 <div class = "container">
     <div class = "panel panel-default">
-		<div class="panel-heading text-center"><h4>Register Your Campus Organization</h4></div>
+		<div class="panel-heading text-center"><h4>Register campus Partner</h4></div>
 			<div class="panel-body">
 
   <!--div class="col-md-offset-2 col-md-8"-->
