+++ partners/templates/registration/community_partner_register_done.html
